Plant harvestcraft crops on sticks with their item
By default all harvestcraft crops can be planted by using the output
item in place of a seed. This behaviour cannot be disabled. Because
these items were not added as valid seeds to agricraft, this allowed
planting these items vanilla-style, which is not what we want.

This adds these items as valid seeds to agricraft, allowing you to plant
them on sticks and preventing you from planting them normally.

Fixes #2.