Add vis seeds to agricraft
The vis seeds (and the plants that grow from these) don't have a
separate texture for each aspect. Instead the images that are colored
based on the aspect are built up of two layers, one that is the same
regardless of the aspect, and one that is recolored based on the aspect
and overlaid on the first. It uses a custom renderer for this, but
unfortunately we cannot use this renderer when using the crops in
AgriCraft.

To work around this this includes a build script that takes the
aforementioned textures + a list of aspects and their colors, and builds
textures (and plant jsons) for all of the available varieties.

To throw an extra wrench in things there is one aspect (caeles) that
does not have a static color but instead cycles through colors. This one
has been special-cased and gets an animated texture.

Part of #1.